Refactor block meta iteration
