+++ pkg/storegateway/bucket.go
@@ -2,7 +2,6 @@ package storegateway
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"path"
 	"path/filepath"
@@ -17,8 +16,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/common/model"
 	"github.com/samber/lo"
-	"github.com/thanos-io/objstore"
-	"golang.org/x/sync/errgroup"
 
 	phlareobj "github.com/grafana/phlare/pkg/objstore"
 	"github.com/grafana/phlare/pkg/phlaredb"
@@ -258,48 +255,25 @@ func (s *BucketStore) fetchBlocksMeta(ctx context.Context) (map[ulid.ULID]*block
 		to   = time.Now()
 		from = to.Add(-time.Hour * 24 * 31) // todo make this configurable
 	)
-	g, ctx := errgroup.WithContext(ctx)
-	g.SetLimit(128)
 
-	allIDs, err := s.listAllBlockByPrefixes(ctx, from, to)
-	if err != nil {
-		return nil, err
-	}
-	totalMeta := 0
-	for _, ids := range allIDs {
-		totalMeta += len(ids)
-	}
-	metas := make([]*block.Meta, totalMeta)
-	idx := 0
+	var (
+		metas []*block.Meta
+		mtx   sync.Mutex
+	)
+
 	start := time.Now()
-	level.Debug(s.logger).Log("msg", "fetching blocks meta", "total", totalMeta)
+	level.Debug(s.logger).Log("msg", "fetching blocks meta", "from", from, "to", to)
 	defer func() {
-		level.Debug(s.logger).Log("msg", "fetched blocks meta", "total", totalMeta, "elapsed", time.Since(start))
+		level.Debug(s.logger).Log("msg", "fetched blocks meta", "total", len(metas), "elapsed", time.Since(start))
 	}()
-	// fetch all meta.json
-	for _, ids := range allIDs {
-		for _, id := range ids {
-			id := id
-			currentIdx := idx
-			idx++
-			g.Go(func() error {
-				r, err := s.bucket.Get(ctx, id+block.MetaFilename)
-				if err != nil {
-					return err
-				}
+	if err := block.IterBlockMetas(ctx, s.bucket, from, to, func(m *block.Meta) {
+		mtx.Lock()
+		defer mtx.Unlock()
+		metas = append(metas, m)
+	}); err != nil {
+		return nil, errors.Wrap(err, "iter block metas")
+	}
 
-				m, err := block.Read(r)
-				if err != nil {
-					return err
-				}
-				metas[currentIdx] = m
-				return nil
-			})
-		}
-	}
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
 	metaMap := lo.SliceToMap(metas, func(item *block.Meta) (ulid.ULID, *block.Meta) {
 		return item.ULID, item
 	})
@@ -314,73 +288,6 @@ func (s *BucketStore) fetchBlocksMeta(ctx context.Context) (map[ulid.ULID]*block
 		}
 	}
 	return metaMap, nil
-}
-
-func (s *BucketStore) listAllBlockByPrefixes(ctx context.Context, from, to time.Time) ([][]string, error) {
-	// todo: We should cache prefixes listing per tenants.
-	blockPrefixes, err := blockPrefixesFromTo(from, to, 4)
-	if err != nil {
-		return nil, err
-	}
-	ids := make([][]string, len(blockPrefixes))
-	g, ctx := errgroup.WithContext(ctx)
-	g.SetLimit(64)
-
-	for i, prefix := range blockPrefixes {
-		prefix := prefix
-		i := i
-		g.Go(func() error {
-			level.Debug(s.logger).Log("msg", "listing blocks", "prefix", prefix, "i", i)
-			prefixIds := []string{}
-			err := s.bucket.Iter(ctx, prefix, func(name string) error {
-				if _, ok := block.IsBlockDir(name); ok {
-					prefixIds = append(prefixIds, name)
-				}
-				return nil
-			}, objstore.WithoutApendingDirDelim)
-			if err != nil {
-				return err
-			}
-			ids[i] = prefixIds
-			return nil
-		})
-	}
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-	return ids, nil
-}
-
-// orderOfSplit is the number of bytes of the ulid id used for the split. The duration of the split is:
-// 0: 1114y
-// 1: 34.8y
-// 2: 1y
-// 3: 12.4d
-// 4: 9h19m
-// TODO: To needs to be adapted based on the MaxBlockDuration.
-func blockPrefixesFromTo(from, to time.Time, orderOfSplit uint8) (prefixes []string, err error) {
-	var id ulid.ULID
-
-	if orderOfSplit > 9 {
-		return nil, fmt.Errorf("order of split must be between 0 and 9")
-	}
-
-	byteShift := (9 - orderOfSplit) * 5
-
-	ms := uint64(from.UnixMilli()) >> byteShift
-	ms = ms << byteShift
-	for ms <= uint64(to.UnixMilli()) {
-		if err := id.SetTime(ms); err != nil {
-			return nil, err
-		}
-		prefixes = append(prefixes, id.String()[:orderOfSplit+1])
-
-		ms = ms >> byteShift
-		ms += 1
-		ms = ms << byteShift
-	}
-
-	return prefixes, nil
 }
 
 // bucketBlockSet holds all blocks.
